+++ examples/wrapper-types/tests/test_generated_bindings.rs
@@ -1,4 +1,8 @@
 uniffi_macros::build_foreign_language_testcases!(
     ["src/wrapper-types.udl"],
-    ["tests/bindings/test_wrapper_types.py"]
+    [
+        "tests/bindings/test_wrapper_types.kts",
+        "tests/bindings/test_wrapper_types.py",
+        "tests/bindings/test_wrapper_types.swift",
+    ]
 );